Moves consumers into a dedicated package
