+++ src/main/scala/it/unibo/authsim/library/dsl/attack/builders/BruteForceAttackBuilder.scala
@@ -1,7 +1,8 @@
 package it.unibo.authsim.library.dsl.attack.builders
 
-import it.unibo.authsim.library.dsl.{HashFunction, Proxy, StatisticsConsumer}
+import it.unibo.authsim.library.dsl.{HashFunction, Proxy}
 import it.unibo.authsim.library.dsl.attack.statistics.Statistics
+import it.unibo.authsim.library.dsl.consumers.StatisticsConsumer
 import it.unibo.authsim.library.user.User
 
 import scala.concurrent.duration.{Duration, MILLISECONDS}
