+++ src/main/scala/it/unibo/authsim/library/dsl/consumers/Consumer.scala
@@ -1,4 +1,4 @@
-package it.unibo.authsim.library.dsl
+package it.unibo.authsim.library.dsl.consumers
 
 trait Consumer[T]:
   def consume(consumable: T): Unit
